requirements.txt
Package                  Version
------------------------ ---------
cachetools               5.2.0
certifi                  2022.9.24
charset-normalizer       2.1.1
google-api-core          2.10.2
google-api-python-client 2.65.0
google-auth              2.13.0
google-auth-httplib2     0.1.0
google-auth-oauthlib     0.7.0
google-cloud-language    2.6.1
googleapis-common-protos 1.56.4
grpcio                   1.50.0
grpcio-status            1.50.0
httplib2                 0.20.4
idna                     3.4
oauthlib                 3.2.2
pip                      22.3
proto-plus               1.22.1
protobuf                 4.21.9
pyasn1                   0.4.8
pyasn1-modules           0.2.8
pyparsing                3.0.9
requests                 2.28.1
requests-oauthlib        1.3.1
rsa                      4.9
setuptools               65.4.1
six                      1.16.0
uritemplate              4.1.1
urllib3                  1.26.12
